+++ libs/qsynedit/qsynedit/syntaxer/glsl.h
@@ -196,6 +196,12 @@ public:
     // Highlighter interface
 public:
     bool supportBraceLevel() override;
+
+    // Syntaxer interface
+public:
+    QString commentSymbol() override;
+    QString blockCommentBeginSymbol() override;
+    QString blockCommentEndSymbol() override;
 };
 
 }