+++ libs/qsynedit/qsynedit/syntaxer/makefile.cpp
@@ -691,4 +691,9 @@ QSet<QString> MakefileSyntaxer::keywords()
     return Directives;
 }
 
+QString MakefileSyntaxer::commentSymbol()
+{
+    return "#";
+}
+
 }